+++ jest.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  moduleNameMapper: {
+    '\\.(scss|css)$': '<rootDir>/__mocks__/styleMock.js',
+  },
+  testMatch: [
+    '**/__tests__/**/*.test.(ts|tsx|js|jsx)',
+    '**/?(*.)+(spec|test).(ts|tsx|js|jsx)'
+  ],
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+};