+++ src/redux/store.ts
@@ -1,5 +1,7 @@
+// DEBUG: Store initialization
+console.log("[DEBUG] Initializing Redux store");
 import { configureStore } from '@reduxjs/toolkit';
-import portfolioReducer from '@redux/portfolioSlice';
+import portfolioReducer from './portfolioSlice';
 import marketDataReducer from '@redux/marketDataSlice';
 import tradingReducer from '@redux/tradingSlice';
 import learningReducer from '@redux/learningSlice';